Modulo 1 - Bases de Datos/MP2/requirements.txt
mysql-connector-python
prettytable
detect-delimiter
chardet
